+++ messaging/messaging-jms/src/main/java/uk/gov/justice/services/messaging/jms/JmsQueueBrowserUnsupportedOperation.java
@@ -1,0 +1,8 @@
+package uk.gov.justice.services.messaging.jms;
+
+public class JmsQueueBrowserUnsupportedOperation extends RuntimeException {
+
+    public JmsQueueBrowserUnsupportedOperation(final String message) {
+        super(message);
+    }
+}