+++ CHANGELOG.md
@@ -5,6 +5,9 @@ on [Keep a CHANGELOG](http://keepachangelog.com/). This project adheres to
 
 ## [Unreleased]
 
+## Added
+- JmsQueueBrowser service to return the size of a queue
+
 ### [6.0.0-M28] - 2019-06-05
 ## Added
 - beans.xml to framework-management module